[MM-40214] Register zoom app bar component (#246) (#261)
