+++ webapp/src/index.js
@@ -3,43 +3,40 @@
 
 import React from 'react';
 
-import {getConfig} from 'mattermost-redux/selectors/entities/general';
-
 import {id as pluginId} from './manifest';
 
-import Icon from './components/icon';
+import ChannelHeaderIcon from './components/channel-header-icon';
 import PostTypeZoom from './components/post_type_zoom';
 import {startMeeting} from './actions';
 import Client from './client';
+import {getPluginURL, getServerRoute} from './selectors';
 
 class Plugin {
     // eslint-disable-next-line no-unused-vars
     initialize(registry, store) {
         registry.registerChannelHeaderButtonAction(
-            <Icon/>,
+            <ChannelHeaderIcon/>,
             (channel) => {
                 startMeeting(channel.id)(store.dispatch, store.getState);
             },
             'Start Zoom Meeting',
+            'Start Zoom Meeting',
         );
+
+        if (registry.registerAppBarComponent) {
+            const iconURL = getPluginURL(store.getState()) + '/public/app-bar-icon.png';
+            registry.registerAppBarComponent(
+                iconURL,
+                (channel) => {
+                    startMeeting(channel.id)(store.dispatch, store.getState);
+                },
+                'Start Zoom Meeting',
+            );
+        }
+
         registry.registerPostTypeComponent('custom_zoom', PostTypeZoom);
         Client.setServerRoute(getServerRoute(store.getState()));
     }
 }
 
 window.registerPlugin(pluginId, new Plugin());
-
-const getServerRoute = (state) => {
-    const config = getConfig(state);
-
-    let basePath = '';
-    if (config && config.SiteURL) {
-        basePath = new URL(config.SiteURL).pathname;
-
-        if (basePath && basePath[basePath.length - 1] === '/') {
-            basePath = basePath.substr(0, basePath.length - 1);
-        }
-    }
-
-    return basePath;
-};
